Refactor: pages폴더 생성
